refactor: Move classes to another package due mongo-db integration.
