--- src/main/java/com/udacity/course3/reviews/entity/Comment.java
@@ -1,92 +1,0 @@
-package com.udacity.course3.reviews.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-@Entity
-@Table(name="comment")
-public class Comment {
-
-  
-  /* member variables */
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name="comment_id")
-  private Integer id;
-  
-  @Column(name="name")
-  @NotNull
-  private String name;
-  
-  @Column(name="content")
-  @NotNull
-  private String content;
-  
-  @Column(name = "created_time")
-  private Timestamp createdTime;
-  
-  @ManyToOne
-  @JoinColumn(name="review_id")
-  @JsonIgnore
-  private Review    review;
-
-  
-  /* constructors */
-  public Comment() {
-    super();
-  }
-
-  
-  /* methods */
-  public Integer getId() {
-    return id;
-  }
-
-  public void setId(Integer id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getContent() {
-    return content;
-  }
-
-  public void setContent(String content) {
-    this.content = content;
-  }
-
-  public Timestamp getCreatedTime() {
-    return createdTime;
-  }
-
-  public void setCreatedTime(Timestamp createdTime) {
-    this.createdTime = createdTime;
-  }
-
-  public Review getReview() {
-    return review;
-  }
-
-  public void setReview(Review review) {
-    this.review = review;
-  }
-
-}
